refactor: Removing shared folder
