+++ dpypelines/pipeline/transforms/sdmx/compact/v20/prototype/manifest.json
