+++ dpypelines/pipeline/transforms/sdmx/compact/v20/prototype/v1.py
@@ -3,8 +3,8 @@ import xml.etree.ElementTree as ET
 import pandas as pd
 import xmltodict
 
-from dpypelines.pipeline.shared.transforms.utils import convert, flatten_dict
-from dpypelines.pipeline.shared.transforms.validate_transform_v20 import (
+from dpypelines.pipeline.transforms.utils import convert, flatten_dict
+from dpypelines.pipeline.transforms.validate_transform_v20 import (
     check_columns_of_dataframes_are_unique,
     check_header_info,
     check_header_unpacked,
