+++ dpypelines/pipeline/transforms/sdmx/generate_versions_metadata.py
@@ -3,7 +3,7 @@ import json
 import pandas as pd
 import xmltodict
 
-from dpypelines.pipeline.shared.transforms.utils import pathify, set_key
+from dpypelines.pipeline.transforms.utils import pathify, set_key
 
 
 def generate_versions_metadata(
