+++ dpypelines/pipeline/transforms/sdmx/generic/v21/prototype/v1.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import xmltodict
 
-from dpypelines.pipeline.shared.transforms.utils import flatten_dict
-from dpypelines.pipeline.shared.transforms.validate_transform_v21 import (
+from dpypelines.pipeline.transforms.utils import flatten_dict
+from dpypelines.pipeline.transforms.validate_transform_v21 import (
     check_columns_of_dataframes_are_unique,
     check_header_info,
     check_header_unpacked,
