+++ dpypelines/pipeline/transforms/sdmx/v20.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
-from dpypelines.pipeline.shared.transforms.sdmx.compact.v20.prototype.v1 import (
+from dpypelines.pipeline.transforms.sdmx.compact.v20.prototype.v1 import (
     xmlToCsvSDMX2_0,
 )
-from dpypelines.pipeline.shared.transforms.sdmx.generate_versions_metadata import (
+from dpypelines.pipeline.transforms.sdmx.generate_versions_metadata import (
     generate_versions_metadata,
 )
